Yacine

+++ src/main/java/com/teamname/buildings/workplaces/WorkplaceService.java
@@ -1,5 +1,6 @@
 package com.teamname.buildings.workplaces;
 
+import com.teamname.exceptions.ResourcesNotFoundException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -19,6 +20,9 @@ public class WorkplaceService{
     }
 
     public Optional<Workplace> getWorkplaceById(int id) {
+        if(workplaceDAO.selectWorkplaceById(id).isEmpty()){
+            throw new ResourcesNotFoundException("Workplace with id: "+ id + "is not found");
+        }
         return workplaceDAO.selectWorkplaceById(id);
     }
 
@@ -27,10 +31,16 @@ public class WorkplaceService{
     }
 
     public void deleteWorkplace(int id){
+        if(workplaceDAO.selectWorkplaceById(id).isEmpty()){
+            throw new ResourcesNotFoundException("Workplace with id: "+ id + "is not found");
+        }
         workplaceDAO.deleteWorkplace(id);
     }
 
     public void updateWorkplace(int id, Workplace workplace){
+        if(workplaceDAO.selectWorkplaceById(id).isEmpty()){
+            throw new ResourcesNotFoundException("Workplace with id: "+ id + "is not found");
+        }
         workplaceDAO.updateWorkplace(id, workplace);
     }
